refac(cspr): minor cleanup, renames and fixes
