+++ libs/ledger-live-common/src/families/casper/bridge.integration.test.ts
@@ -12,7 +12,7 @@ import {
 } from "@ledgerhq/errors";
 
 import { CasperInvalidTransferId } from "./errors";
-import { getEstimatedFees } from "./bridge/utils/fee";
+import { getEstimatedFees } from "./bridge/bridgeHelpers/fee";
 
 const SEED_IDENTIFIER = "0202ba6dc98cbe677711a45bf028a03646f9e588996eb223fad2485e8bc391b01581";
 const ACCOUNT_2 = "0203A17118eC0e64c4e4FdbDbEe0eA14D118C9aAf08C6c81bbB776Cae607cEB84EcB";
