+++ libs/ledger-live-common/src/families/casper/bridge/bridgeHelpers/fee.ts
