+++ libs/ledger-live-common/src/families/casper/bridge/bridgeHelpers/transferId.ts
