+++ libs/ledger-live-common/src/families/casper/bridge/bridgeHelpers/types.ts
