+++ libs/ledger-live-common/src/families/casper/errors.ts
@@ -1,3 +1,5 @@
 import { createCustomErrorClass } from "@ledgerhq/errors";
 
 export const CasperInvalidTransferId = createCustomErrorClass("CasperInvalidTransferId");
+export const InvalidMinimumAmount = createCustomErrorClass("InvalidMinimumAmount");
+export const MayBlockAccount = createCustomErrorClass("MayBlockAccount");
