+++ libs/ledgerjs/packages/errors/src/index.ts
@@ -11,8 +11,6 @@ export { serializeError, deserializeError, createCustomErrorClass, addCustomErro
 export const AccountNameRequiredError = createCustomErrorClass("AccountNameRequired");
 export const AccountNotSupported = createCustomErrorClass("AccountNotSupported");
 export const AmountRequired = createCustomErrorClass("AmountRequired");
-export const InvalidMinimumAmount = createCustomErrorClass("InvalidMinimumAmount");
-export const MayBlockAccount = createCustomErrorClass("MayBlockAccount");
 export const BluetoothRequired = createCustomErrorClass("BluetoothRequired");
 export const BtcUnmatchedApp = createCustomErrorClass("BtcUnmatchedApp");
 export const CantOpenDevice = createCustomErrorClass("CantOpenDevice");
@@ -79,7 +77,6 @@ export const ManagerUninstallBTCDep = createCustomErrorClass("ManagerUninstallBT
 export const NetworkDown = createCustomErrorClass("NetworkDown");
 export const NoAddressesFound = createCustomErrorClass("NoAddressesFound");
 export const NotEnoughBalance = createCustomErrorClass("NotEnoughBalance");
-export const InvalidAmountTransfer = createCustomErrorClass("InvalidAmountTransfer");
 export const NotEnoughBalanceToDelegate = createCustomErrorClass("NotEnoughBalanceToDelegate");
 export const NotEnoughBalanceInParentAccount = createCustomErrorClass(
   "NotEnoughBalanceInParentAccount",
